// src/moveAbortCatalog.generated.ts
export const MOVE_ABORT_CATALOG: Record<
  string,
  Record<number, MoveAbortEntry>
> = {
// ...
  "trading_account": {
    "0": {
      "name": "EInvalidOwner",
      "label": "Invalid owner",
      "message": null,
      "clever": false,
      "source": "packages/triex/sources/trading_account.move:26"
    },
    "1": {
      "name": "EInvalidTrader",
      "label": "Invalid trader",
      "message": null,
      "clever": false,
      "source": "packages/triex/sources/trading_account.move:27"
    },
    "2": {
      "name": "EInvalidProof",
      "label": "Invalid proof",
      "message": null,
      "clever": false,
      "source": "packages/triex/sources/trading_account.move:28"
    },
    "3": {
      "name": "ETradingAccountBalanceTooLow",
      "label": "Trading account balance too low",
      "message": null,
      "clever": false,
      "source": "packages/triex/sources/trading_account.move:29"
    },
    "4": {
      "name": "EMaxCapsReached",
      "label": "Max caps reached",
      "message": null,
      "clever": false,
      "source": "packages/triex/sources/trading_account.move:30"
    },
    "5": {
      "name": "ECapNotInList",
      "label": "Cap not in list",
      "message": null,
      "clever": false,
      "source": "packages/triex/sources/trading_account.move:31"
    },
    "7": {
      "name": "EMultiCoinBalanceTooLow",
      "label": "Multi coin balance too low",
      "message": null,
      "clever": false,
      "source": "packages/triex/sources/trading_account.move:34"
    }
  },
// ...
}